+++ packages/data-ferret/src/lib/locateCircularReference/index.ts
@@ -1,0 +1,1 @@
+export { locateCircularReference } from './locateCircularReference'