+++ packages/data-ferret/src/lib/models/index.ts
@@ -1,1 +1,2 @@
 export type { Predicate, DataType } from '../shared/model'
+export { CircularReference } from '../CircularReference'